Pulled out all the basic data into the constants.json file

+++ Application/Salvation.Core/Modelling/HolyPriest/Spells/SpellService.cs
@@ -154,7 +154,7 @@ namespace Salvation.Core.Modelling.HolyPriest.Spells
 
             var baseMana = _gameStateService.GetBaseManaAmount(gameState);
 
-            return baseMana * spellData.ManaCost;
+            return baseMana * (decimal)spellData.ManaCost;
         }
 
         public virtual decimal GetDuration(GameState gameState, BaseSpellData spellData = null,
